Implement pic messaging
