+++ src/main/kotlin/db/Db.kt
@@ -1,17 +1,73 @@
 package com.neelkamath.omniChat.db
 
 import com.neelkamath.omniChat.*
+import com.neelkamath.omniChat.db.Pic.Companion.MAX_BYTES
 import com.neelkamath.omniChat.db.tables.*
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.like
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.postgresql.util.PGobject
+import javax.annotation.Generated
 
 data class ChatEdges(val chatId: Int, val edges: List<MessageEdge>)
 
 data class ForwardPagination(val first: Int? = null, val after: Int? = null)
 
 data class BackwardPagination(val last: Int? = null, val before: Int? = null)
+
+/** Throws an [IllegalArgumentException] if the [bytes] exceeds [Pic.MAX_BYTES]. */
+data class Pic(
+    /** At most [MAX_BYTES]. */
+    val bytes: ByteArray,
+    val type: Type
+) {
+    init {
+        if (bytes.size > MAX_BYTES)
+            throw IllegalArgumentException("The pic mustn't exceed $MAX_BYTES bytes.")
+    }
+
+    /** @see [buildType] */
+    enum class Type {
+        PNG {
+            override fun toString() = "png"
+        },
+        JPEG {
+            override fun toString() = "jpg"
+        }
+    }
+
+    @Generated
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        other as Pic
+
+        if (!bytes.contentEquals(other.bytes)) return false
+        if (type != other.type) return false
+
+        return true
+    }
+
+    @Generated
+    override fun hashCode(): Int {
+        var result = bytes.contentHashCode()
+        result = 31 * result + type.hashCode()
+        return result
+    }
+
+    companion object {
+        const val MAX_BYTES = 25 * 1024 * 1024
+
+        /** Throws an [IllegalArgumentException] if the [extension] (e.g., `"pjpeg"`) isn't one of the [Type]s. */
+        fun buildType(extension: String): Type = when (extension) {
+            "png" -> Type.PNG
+            "jpg", "jpeg", "jfif", "pjpeg", "pjp" -> Type.JPEG
+            else ->
+                throw IllegalArgumentException("The pic ($extension) must be one of ${Type.values().joinToString()}.")
+        }
+    }
+}
 
 /**
  * Required for enums (see https://github.com/JetBrains/Exposed/wiki/DataTypes#how-to-use-database-enum-types). It's
@@ -49,6 +105,7 @@ private fun create(): Unit = transaction {
     createTypes()
     SchemaUtils.create(
         TextMessages,
+        PicMessages,
         AudioMessages,
         Pics,
         Contacts,
