+++ src/main/kotlin/db/tables/Users.kt
@@ -20,16 +20,13 @@ data class User(
     val pic: Pic?
 )
 
-/** Replacement for Keycloak's cumbersome custom user attributes. Pics cannot exceed [Pics.MAX_PIC_BYTES]. */
+/** Replacement for Keycloak's cumbersome custom user attributes. Pics cannot exceed [Pic.MAX_BYTES]. */
 object Users : IntIdTable() {
     /** The ID given to the user by Keycloak. */
     private val uuid: Column<UUID> = uuid("uuid")
 
-    /** Bios cannot exceed this many characters. */
-    const val MAX_BIO_LENGTH = 2500
-
     private val picId: Column<Int?> = integer("pic_id").references(Pics.id).nullable()
-    private val bio: Column<String?> = varchar("bio", MAX_BIO_LENGTH).nullable()
+    private val bio: Column<String?> = varchar("bio", Bio.MAX_LENGTH).nullable()
     private val isOnline: Column<Boolean> = bool("is_online").clientDefault { false }
     private val lastOnline: Column<LocalDateTime?> = datetime("last_online").nullable()
