+++ src/main/kotlin/restApi/HealthCheck.kt
@@ -1,0 +1,11 @@
+package com.neelkamath.omniChat.restApi
+
+import io.ktor.application.call
+import io.ktor.http.HttpStatusCode
+import io.ktor.response.respond
+import io.ktor.routing.Routing
+import io.ktor.routing.get
+
+fun routeHealthCheck(routing: Routing): Unit = with(routing) {
+    get("health-check") { call.respond(HttpStatusCode.NoContent) }
+}
