+++ src/test/kotlin/ApiModelsTest.kt
@@ -1,8 +1,5 @@
 package com.neelkamath.omniChat
 
-import com.neelkamath.omniChat.db.tables.GroupChats
-import com.neelkamath.omniChat.db.tables.TextMessages
-import com.neelkamath.omniChat.db.tables.Users
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
 
@@ -38,7 +35,7 @@ class GroupChatInputTest : FunSpec({
 class BioTest : FunSpec({
     context("init") {
         test("An exception should be thrown if the value is too big") {
-            val value = CharArray(Users.MAX_BIO_LENGTH + 1) { 'a' }.joinToString("")
+            val value = CharArray(Bio.MAX_LENGTH + 1) { 'a' }.joinToString("")
             shouldThrowExactly<IllegalArgumentException> { Bio(value) }
         }
     }
@@ -55,7 +52,7 @@ class TextMessageTest : FunSpec({
         }
 
         test("An exception should be thrown if the value is too long") {
-            val text = CharArray(TextMessages.MAX_TEXT_LENGTH + 1) { 'a' }.joinToString("")
+            val text = CharArray(TextMessage.MAX_LENGTH + 1) { 'a' }.joinToString("")
             shouldThrowExactly<IllegalArgumentException> { TextMessage(text) }
         }
     }
@@ -68,7 +65,7 @@ class GroupChatTitleTest : FunSpec({
         }
 
         test("An exception should be thrown if the title is too long") {
-            val title = CharArray(GroupChats.MAX_TITLE_LENGTH + 1) { 'a' }.joinToString("")
+            val title = CharArray(GroupChatTitle.MAX_LENGTH + 1) { 'a' }.joinToString("")
             shouldThrowExactly<IllegalArgumentException> { GroupChatTitle(title) }
         }
 
@@ -81,7 +78,7 @@ class GroupChatTitleTest : FunSpec({
 class GroupChatDescriptionTest : FunSpec({
     context("init") {
         test("An exception should be thrown if the description is too long") {
-            val description = CharArray(GroupChats.MAX_DESCRIPTION_LENGTH + 1) { 'a' }.joinToString("")
+            val description = CharArray(GroupChatDescription.MAX_LENGTH + 1) { 'a' }.joinToString("")
             shouldThrowExactly<IllegalArgumentException> { GroupChatDescription(description) }
         }
     }
