+++ src/test/kotlin/AppTest.kt
@@ -7,7 +7,7 @@ import com.neelkamath.omniChat.db.tables.create
 import com.neelkamath.omniChat.graphql.operations.READ_ACCOUNT_QUERY
 import com.neelkamath.omniChat.graphql.operations.REQUEST_TOKEN_SET_QUERY
 import com.neelkamath.omniChat.graphql.operations.createMessage
-import com.neelkamath.omniChat.routing.readGraphQlHttpResponse
+import com.neelkamath.omniChat.graphql.routing.readGraphQlHttpResponse
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.maps.shouldContain
 import io.kotest.matchers.maps.shouldNotHaveKey
@@ -43,12 +43,18 @@ class EncodingTest : FunSpec({
 class SpecComplianceTest : FunSpec({
     test("""The "data" key shouldn't be returned if there was no data to be received""") {
         val login = Login(Username("username"), Password("password"))
-        readGraphQlHttpResponse(REQUEST_TOKEN_SET_QUERY, variables = mapOf("login" to login)) shouldNotHaveKey "data"
+        readGraphQlHttpResponse(
+            REQUEST_TOKEN_SET_QUERY,
+            variables = mapOf("login" to login)
+        ) shouldNotHaveKey "data"
     }
 
     test("""The "errors" key shouldn't be returned if there were no errors""") {
         val login = createVerifiedUsers(1)[0].login
-        readGraphQlHttpResponse(REQUEST_TOKEN_SET_QUERY, variables = mapOf("login" to login)) shouldNotHaveKey
+        readGraphQlHttpResponse(
+            REQUEST_TOKEN_SET_QUERY,
+            variables = mapOf("login" to login)
+        ) shouldNotHaveKey
                 "errors"
     }
 
@@ -57,7 +63,10 @@ class SpecComplianceTest : FunSpec({
         createUser(account)
         val userId = readUserByUsername(account.username).id
         val accessToken = buildAuthToken(userId).accessToken
-        val response = readGraphQlHttpResponse(READ_ACCOUNT_QUERY, accessToken = accessToken)["data"] as Map<*, *>
+        val response = readGraphQlHttpResponse(
+            READ_ACCOUNT_QUERY,
+            accessToken = accessToken
+        )["data"] as Map<*, *>
         objectMapper.convertValue<Map<String, Any>>(response["readAccount"]!!) shouldContain Pair("firstName", null)
     }
 })
