+++ src/test/kotlin/db/tables/AudioMessagesTest.kt
@@ -6,7 +6,7 @@ import io.kotest.core.spec.style.FunSpec
 class Mp3Test : FunSpec({
     context("init") {
         test("An excessively large audio file should cause an exception to be thrown") {
-            shouldThrowExactly<IllegalArgumentException> { Mp3(ByteArray(AudioMessages.MAX_AUDIO_BYTES + 1)) }
+            shouldThrowExactly<IllegalArgumentException> { Mp3(ByteArray(Mp3.MAX_BYTES + 1)) }
         }
     }
 })
