+++ src/test/kotlin/graphql/engine/MachineryUtil.kt
@@ -2,8 +2,8 @@ package com.neelkamath.omniChat.graphql.engine
 
 import com.fasterxml.jackson.module.kotlin.convertValue
 import com.neelkamath.omniChat.GraphQlResponse
+import com.neelkamath.omniChat.graphql.routing.readGraphQlHttpResponse
 import com.neelkamath.omniChat.objectMapper
-import com.neelkamath.omniChat.routing.readGraphQlHttpResponse
 import graphql.ExecutionInput
 
 /**
