+++ src/test/kotlin/graphql/routing/QueriesAndMutationsUtil.kt
@@ -1,4 +1,4 @@
-package com.neelkamath.omniChat.routing
+package com.neelkamath.omniChat.graphql.routing
 
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.neelkamath.omniChat.GraphQlRequest
